Move serviceendpoint start functionality to register function
